Started work on particles
Added a new pipeline, subpass, framebuffer attachment, and shaders.

The first subpass in the deferred renderpass is the old renderpass (which
does not render particles).  The second subpass renders particles only
(but still does a depth test) to a different output attachment.  The particle
attachment is blended into the scene after lighting is done.  The particles
have support for billboarding (only on every axis currently).  Hybrid rendermode
is the only supported method, and shadows currently do not work properly.
Raytracing will need a custom intersection shader in order for the billboards to always
face the incoming ray (instead of always facing the camera).

Some of the FFXI generator is parsed, but most of it is unused still.  A temporary
testing file was written in order to set up the renderer section.